+++ packages/nitro/src/vite.ts
@@ -1,8 +1,8 @@
+import type { Nitro } from 'nitro/types';
 import type { Plugin } from 'vite';
 import type { ModuleOptions } from './index.js';
-import type { Nitro } from 'nitro/types';
-import { workflowRollupPlugin } from './rollup-plugin.js';
 import nitroModule from './index.js';
+import { workflowRollupPlugin } from './rollup.js';
 
 export function workflow(options?: ModuleOptions): Plugin[] {
   return [
@@ -10,7 +10,7 @@ export function workflow(options?: ModuleOptions): Plugin[] {
     {
       name: 'workflow:nitro',
       // Requires https://github.com/nitrojs/nitro/discussions/3680
-      // @ts-ignore
+      // @ts-expect-error
       nitro: {
         setup: (nitro: Nitro) => {
           nitro.options.workflow = {